requirements.txt
# API Server dependencies
fastapi>=0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
pydantic>=2.9.0
python-dotenv==1.0.0

# Gemini API
google-genai==1.64.0

# Utilities
requests==2.31.0
python-jose==3.3.0
bcrypt==4.1.2
PyJWT==2.8.0
